requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
ollama>=0.2.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for advanced features
rich>=13.0.0  # Enhanced terminal output
click>=8.1.0  # CLI interface (if needed later)